+++ admodels/rtb_source.go
@@ -67,10 +67,10 @@ func RTBSourceFromModel(cl *models.RTBSource, acc *Account) (src *RTBSource) {
 	}
 
 	filter := types.BaseFilter{
-		Secure:          cl.Secure,
-		Adblock:         cl.AdBlock,
-		PrivateBrowsing: cl.PrivateBrowsing,
-		IP:              cl.IP,
+		Secure:          int8(cl.Secure),
+		Adblock:         int8(cl.AdBlock),
+		PrivateBrowsing: int8(cl.PrivateBrowsing),
+		IP:              int8(cl.IP),
 	}
 
 	filter.Set(types.FieldFormat, cl.Formats)
